refactor: rename files and update references for inverted pendulum an… (#1171)
* refactor: rename files and update references for inverted pendulum and path tracking modules

* refactor: rename inverted pendulum control files and update type check references

* refactor: update import statements to use consistent casing for InvertedPendulum module
