+++ docs/index_main.rst
@@ -43,7 +43,7 @@ this graph shows GitHub star history of this project:
    modules/7_arm_navigation/arm_navigation
    modules/8_aerial_navigation/aerial_navigation
    modules/9_bipedal/bipedal
-   modules/10_control/control
+   modules/10_inverted_pendulum/inverted_pendulum
    modules/11_utils/utils
    modules/12_appendix/appendix
