+++ docs/modules/10_inverted_pendulum/inverted-pendulum.png
