+++ docs/modules/11_utils/utils_main.rst
@@ -1,4 +1,4 @@
-.. _utils:
+.. _`utils`:
 
 Utilities
 ==========
