+++ docs/modules/1_introduction/3_technologies_for_robotics/technologies_for_robotics_main.rst
@@ -56,3 +56,13 @@ Robot type specific technologies
 
 #. :ref:`Aerial Navigation`
 #. :ref:`Bipedal`
+#. :ref:`Inverted Pendulum`
+
+
+Additional Information
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+#. :ref:`utils`
+#. :ref:`Appendix`
+
+
