+++ docs/modules/6_path_tracking/move_to_a_pose_control/move_to_a_pose_control_main.rst
