+++ docs/modules/6_path_tracking/path_tracking_main.rst
@@ -16,3 +16,4 @@ Path tracking is the ability of a robot to follow the reference path generated b
    lqr_speed_and_steering_control/lqr_speed_and_steering_control
    model_predictive_speed_and_steering_control/model_predictive_speed_and_steering_control
    cgmres_nmpc/cgmres_nmpc
+   move_to_a_pose_control/move_to_a_pose_control
