+++ tests/test_move_to_pose_robot.py
@@ -1,5 +1,5 @@
 import conftest  # Add root path to sys.path
-from Control.move_to_pose import move_to_pose as m
+from PathTracking.move_to_pose import move_to_pose as m
 
 
 def test_1():
